// src/utils/types.d.ts
export declare namespace Paradise {
  interface MovieCardInfo {
    id: string;
    overview: string;
    title: string;
    poster_path: string;
    backdrop_path: string;
    genres: string;
    release_date: string;
    vote_average: string;
    production_companies: string;
    runtime: string;
    popularity: number;
  }
}